Now reusing TripleColonExtension.
