--- src/Microsoft.DocAsCode.MarkdigEngine.Extensions/Image/ImageExtension.cs
@@ -1,40 +1,0 @@
-// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-namespace Microsoft.DocAsCode.MarkdigEngine.Extensions
-{
-    using Markdig;
-    using Markdig.Extensions.CustomContainers;
-    using Markdig.Renderers;
-
-    public class ImageExtension : IMarkdownExtension
-    {
-        private readonly MarkdownContext _context;
-
-        public ImageExtension(MarkdownContext context)
-        {
-            _context = context;
-        }
-
-        public void Setup(MarkdownPipelineBuilder pipeline)
-        {
-            if (pipeline.BlockParsers.Contains<CustomContainerParser>())
-            {
-                pipeline.BlockParsers.InsertBefore<CustomContainerParser>(new ImageParser(_context));
-            }
-            else
-            {
-                pipeline.BlockParsers.AddIfNotAlready(new ImageParser(_context));
-            }            
-        }
-
-        public void Setup(MarkdownPipeline pipeline, IMarkdownRenderer renderer)
-        {
-            var htmlRenderer = renderer as HtmlRenderer;
-            if (htmlRenderer != null && !htmlRenderer.ObjectRenderers.Contains<ImageRender>())
-            {
-                htmlRenderer.ObjectRenderers.Insert(0, new ImageRender());
-            }
-        }
-    }
-}
