+++ src/Microsoft.DocAsCode.MarkdigEngine.Extensions/TripleColon/ZoneExtension.cs
@@ -2,7 +2,8 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 namespace Microsoft.DocAsCode.MarkdigEngine.Extensions
 {
-	using Markdig.Renderers;
+    using Markdig.Parsers;
+    using Markdig.Renderers;
 	using Markdig.Renderers.Html;
 	using Markdig.Syntax;
 	using System;
@@ -16,13 +17,14 @@ namespace Microsoft.DocAsCode.MarkdigEngine.Extensions
         private static readonly Regex pivotReplaceCommasRegex = new Regex(@"\s*,\s*");
         public string Name => "zone";
 		public bool SelfClosing => false;
+        public bool EndingTripleColons => false;
 
-		public bool Render(HtmlRenderer renderer, TripleColonBlock block)
+        public bool Render(HtmlRenderer renderer, TripleColonBlock block)
 		{
 			return false;
 		}
 
-		public bool TryProcessAttributes(IDictionary<string, string> attributes, out HtmlAttributes htmlAttributes, out IDictionary<string, string> renderProperties, Action<string> logError)
+		public bool TryProcessAttributes(IDictionary<string, string> attributes, out HtmlAttributes htmlAttributes, out IDictionary<string, string> renderProperties, Action<string> logError, BlockProcessor processor)
         {
             htmlAttributes = null;
 			renderProperties = null;
